chore(validations): Aligning config validations across endpoints
